Remove generators hooks, fix bugs
For the time being, hooks can't change the type, so they are of little
interest expect to get data consistent. I'll see that later. However,
of course you can overrides inferred specs.

Something better would be to make enum generated values instances of
the enum, not mere strings.

Also, I should add flag in the context so I can choose to be recursive
or not, and to erase already defined specs or not.